+++ terra-rust-api/Changelog.md
@@ -3,6 +3,7 @@
 ## 1.2
 ### 1.2.1 - 31-Jan-22
 * TX block doesn't require contract/execute_msg parts when it is returned (backwards incompatible)
+* MsgMigrateContract message added
 ## 1.1
 ### 1.1.8 - 17-Dec-21
 * extra fields in SyncInfo area in RPC